cash-track-backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
reportlab==4.2.5
openpyxl==3.1.5
Pillow==11.0.0

